Stop building System.Threading.Tasks.Extensions package
Fixes build breaks after shared CoreLib updates
