--- src/System.Threading.Tasks.Extensions/pkg/System.Threading.Tasks.Extensions.pkgproj
@@ -1,28 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), Directory.Build.props))\Directory.Build.props" />
-  <PropertyGroup>
-    <!-- we need to be supported on pre-nuget-3 platforms (Dev12, Dev11, etc) -->
-    <MinClientVersion>2.8.6</MinClientVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\ref\System.Threading.Tasks.Extensions.csproj">
-      <SupportedFramework>net45</SupportedFramework>
-    </ProjectReference>
-    <ProjectReference Include="..\src\System.Threading.Tasks.Extensions.csproj" />
-
-    <InboxOnTargetFramework Include="netcoreapp2.1" />
-    <InboxOnTargetFramework Include="$(UAPvNextTFM)" />
-    <InboxOnTargetFramework Include="$(AllXamarinFrameworks)" />
-
-    <!-- this package is part of the implementation closure of NETStandard.Library
-         therefore it cannot reference NETStandard.Library -->
-    <SuppressMetaPackage Include="NETStandard.Library" />
-  </ItemGroup>
-  <PropertyGroup>
-    <!-- Excluding the reference assets on the package so that RAR will see the run-time conflicts at build time in order to
-    generate the right binding redirects when targeting Desktop. https://github.com/dotnet/corefx/issues/32457 -->
-    <ExcludeReferenceAssets>true</ExcludeReferenceAssets>
-  </PropertyGroup>
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), Directory.Build.targets))\Directory.Build.targets" />
-</Project>
