+++ src/System.Threading.Tasks.Extensions/ref/Configurations.props
@@ -1,13 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
-    <PackageConfigurations>
-      netstandard1.0;
-      netstandard;
-    </PackageConfigurations>
     <BuildConfigurations>
-      $(PackageConfigurations);
-      _netcoreapp2.1;
       netcoreapp;
       uap;
     </BuildConfigurations>
