+++ src/System.Threading.Tasks.Extensions/tests/Configurations.props
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <BuildConfigurations>
       netcoreapp;
-      netstandard;
       uap;
     </BuildConfigurations>
   </PropertyGroup>
